requirements.txt
groq
streamlit
pandas
azure-ai-formrecognizer
json